Added Service class for basic bus interactions
The service class is similar to the broker class in the last commit as
it has its own event loop and opens a router socket. It differs in the
call backs it uses when receiving a message.

Might possibly make a new base service class to encapsulate common
features between this and the broker.