Added Timers to Reactors
The Reactor class now includes a method to add timers. Where as msg_handlers are for incoming events, timers are used for data which needs be distributed. Using the threading Timer class, custom callbacks can now be added similarly to how add_msg_handlers() works and keyword/positional parameters can be usd normally. Converting this to C++ is going to be a pain in the ass....Good luck who every inherits this.